rename user_manager to dependencies
