+++ fixbackend/workspaces/dependencies.py
@@ -1,0 +1,9 @@
+# from fixbackend.ids import WorkspaceId
+# from fixbackend.workspaces.repositories import WorkspaceRepository
+# from fixbackend.auth.dependencies import AuthenticatedUser
+
+
+# def workspace_from_path(
+#     workspace_id: WorkspaceId, user_context: AuthenticatedUser, workspace_repository: WorkspaceRepositoryDependency
+# ) -> Workspace:
+#     return workspace_repository.get(workspace_id)
